requirements.txt
numpy==1.26.2
scipy==1.11.0
Pillow==7.0.0
torch== 1.4.0
torchvision==0.5.0
pandas==2.1.4
opencv-python==4.8.1
python-dateutil==2.8.2